+++ feed/index.html
@@ -42,7 +42,7 @@
                     <a class="nav-link" href="/profile/index.html">Profile</a>
                   </li>
                   <li class="nav-item">
-                    <a class="nav-link" href="/index.html">Logout</a>
+                    <a class="nav-link logoutLink" href="/index.html">Logout</a>
                   </li>
                 </ul>
               </div>
