+++ index.html
@@ -7,8 +7,7 @@
     <meta name="description" content="FetchFriends sign in or register" />
     <title>FetchFriends social media</title>
     <link rel="stylesheet" href="/dist/css/index.css" />
-    <script type="module" src="js/navigation/pills-nav.mjs"></script>
-    <script type="module" src="js/api/login_signup/login-api.mjs"></script>
+    <script type="module" src="js/index.mjs"></script>
     <script
       src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.1/dist/js/bootstrap.bundle.min.js"
       integrity="sha384-HwwvtgBNo3bZJJLYd8oVXjrBZt8cqVSpeBNS5n7C8IVInixGAoxmnlMuBnhbgrkm"
