+++ js/feed.mjs
@@ -2,6 +2,7 @@ import { fetchPosts } from "./api/posts/fetch-posts.mjs";
 import { renderPosts } from "./api/posts/render-posts.mjs";
 import { addFilterEventListener } from "./api/posts/filter-and-search/filter-posts.mjs";
 import { addSearchEventListener } from "./api/posts/filter-and-search/search-posts.mjs";
+import "./api/login_signup/logout.mjs";
 
 // Define viewPost function
 export function viewPost(postId) {
