+++ js/index.mjs
@@ -1,0 +1,3 @@
+import "./api/login_signup/login-api.mjs";
+import "./api/login_signup/register-api.mjs";
+import "./navigation/pills-nav.mjs";
